backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
pgvector
requests
